+++ VulkanMemoryAllocator/changelog.md
@@ -2,6 +2,9 @@
 
 ## WIP
 
+## [0.6] - 2021-06-22
+- Bump VMA, adding alignment info to PoolCreateInfo
+
 ## [0.5.1] - 2021-06-05
 - Bump VMA, documentation changes