+++ changelog.md
@@ -2,6 +2,9 @@
 
 ## WIP
 
+## [3.11] - 2021-06-21
+- Bump API version to v1.2.182
+
 ## [3.10.4] - 2021-06-14
 - Bump API version to v1.2.181